+++ dungeon-game-core/src/com/dungeon/game/entity/Static.java
@@ -3,7 +3,7 @@ package com.dungeon.game.entity;
 import com.dungeon.game.world.World;
 
 public abstract class Static extends Entity {
-	public Static(World world, int x, int y) {
+	public Static(World world, float x, float y) {
 		super(world, x, y);
 	}
 }
