+++ dungeon-game-core/src/com/dungeon/game/entity/hud/HealthBar.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.scenes.scene2d.utils.TiledDrawable;
 import com.dungeon.game.world.World;
 
 public class HealthBar extends StatusBar {
-	public HealthBar(World world, int x, int y){
+	public HealthBar(World world, float x, float y){
 		super(world, x, y);
 		filler = new TiledDrawable(new TextureRegion(new Texture("heart.png")));
 	}
