+++ dungeon-game-core/src/com/dungeon/game/entity/hud/InvGraphic.java
@@ -9,14 +9,14 @@ public class InvGraphic extends Window {
 	
 	public Inventory inv;
 
-	public InvGraphic(World world, Inventory inv, int x, int y) {
+	public InvGraphic(World world, Inventory inv, float x, float y) {
 		super(world, x, y);
 		d_width = 0;
 		d_height = 14;
 		this.inv = inv;
 		for(Slot slot: inv.slot){
-			if(slot.x + 40>d_width)d_width = slot.x + 40;
-			if(slot.y + 50>d_height)d_height = slot.y + 50;
+			if(slot.x + 40>d_width) d_width = (int) (slot.x + 40);
+			if(slot.y + 50>d_height) d_height = (int) (slot.y + 50);
 		}
 	}
